back/requirement.txt
flask
flask-cors
flask-sqlalchemy
python-dotenv
psycopg2-binary
werkzeug
PyJWT
